feat: update remote config values (#695)

+++ lib/shared/notifiers/timer_notifier/timer_notipod.dart
@@ -3,7 +3,7 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'timer_notifier.dart';
 
 final timerNotipod =
-    StateNotifierProvider.autoDispose.family<TimerNotifier, int, double>(
+    StateNotifierProvider.autoDispose.family<TimerNotifier, int, int>(
   (ref, initial) {
     return TimerNotifier(initial);
   },
